Add: sync_to_async to endpoint where it is required and little project refactor
