+++ labs/llm.py
@@ -7,7 +7,7 @@ from embeddings.openai import OpenAIEmbedder
 from embeddings.vectorizers.chunk_vectorizer import ChunkVectorizer
 from litellm_service.local import RequestLocalLLM
 from litellm_service.request import RequestLiteLLM
-from response_parser.parser import is_valid_json, parse_llm_output
+from parsers.response_parser import is_valid_json, parse_llm_output
 
 logger = logging.getLogger(__name__)
